Remove .remove_from_bounce_list in spec
The ported sendgrid_api_spec is nearly 400 lines long and somewhat
confusing. This breaks out the .remove_from_bounce_list method into an
individual spec that is smaller and conceptually cleaner.
